requirements.txt
Django==2.0.2
requests==2.18.4
beautifulsoup4==4.6.0
djangorestframework==3.8.2
